Add files via upload
SRS stand for “software requirement specification”. This is description of overall Applications development. This is a starting point of Application development OR project. In “SRS” document we prepared scope of project. We also create the use case diagram demonstrate the customer of application. We are creating the project plan and discuss the development methodologies in SRS document. SRS leads the abstract idea to its concretion.